refactor: lottoSimulator 기능 분리
