+++ src/controller/LottoSimulator.js
@@ -1,7 +1,5 @@
-import Lotto from '../domain/Lotto';
 import Validator from '../utils/Validator';
-import { ERROR_MESSAGE, LOTTO_CONSTANT, LOTTO_RANKING, WINNING_PRIZE } from '../data/constants';
-import { RandomNumberGenerator } from '../utils/RandomNumberGenerator';
+import { ERROR_MESSAGE, LOTTO_CONSTANT, LOTTO_RANKING } from '../data/constants';
 import InputView from '../view/InputView';
 import OutputView from '../view/OutputView';
 
@@ -36,32 +34,11 @@ class LottoSimulator {
     try {
       this.validateBudget(budget);
       this.budget = budget;
+      // next
     } catch (err) {
       OutputView.printErrorMessage(err);
       this.inputBudget();
     }
-  }
-
-  createLottoNumbers() {
-    const lottoNumbers = new Set();
-    while (lottoNumbers.size < LOTTO_CONSTANT.LENGTH) {
-      lottoNumbers.add(
-        RandomNumberGenerator.generateNumberInRange(
-          LOTTO_CONSTANT.MIN_NUMBER,
-          LOTTO_CONSTANT.MAX_NUMBER
-        )
-      );
-    }
-    return Array.from(lottoNumbers);
-  }
-
-  purchaseLottos(budget) {
-    this.validateBudget(budget);
-    this.#budget = budget;
-    const lottoCount = budget / LOTTO_CONSTANT.PRICE;
-    Array.from({ length: lottoCount }).forEach(() => {
-      this.#lottos.push(new Lotto(this.createLottoNumbers()));
-    });
   }
 
   getLottoCount() {
@@ -85,17 +62,6 @@ class LottoSimulator {
       winningResult[rank] += 1;
     });
     return winningResult;
-  }
-
-  calculateYieldRate() {
-    const winningResult = this.calculateWinningResult();
-
-    const totalPrize = Object.keys(winningResult).reduce(
-      (sum, rank) => sum + WINNING_PRIZE[rank] * winningResult[rank],
-      0
-    );
-
-    return ((totalPrize / this.#budget) * 100).toFixed(1);
   }
 
   reset() {
