--- src/data/Converter.js
@@ -1,12 +1,0 @@
-// const toOrdinalNumber = Object.freeze({
-//   6: 'FIRST',
-//   10005: 'SECOND',
-//   5: 'THIRD',
-//   4: 'FOURTH',
-//   3: 'FIFTH',
-//   2: 'FAIL',
-//   1: 'FAIL',
-//   0: 'FAIL',
-// });
-
-// export { toOrdinalNumber };
